requirements.txt
simpy>=4.0.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
PyYAML>=6.0
rapidfuzz>=3.0.0